+++ neo/tools/guied/guied.cpp
@@ -145,7 +145,7 @@ Returns true if the given string is an expression
 */
 bool IsExpression ( const char* s )
 {
-	idParser src( s, strlen ( s ), "", 
+	idParser src( s, (int)strlen ( s ), "",
 				  LEXFL_ALLOWMULTICHARLITERALS		| 
 				  LEXFL_NOSTRINGCONCAT				| 
 				  LEXFL_ALLOWBACKSLASHSTRINGCONCAT	|
