+++ neo/tools/radiant/LightDlg.cpp
@@ -637,7 +637,7 @@ void CLightDlg::LoadLightTextures() {
 		mat = declManager->MaterialByIndex(i, false);
 		idStr str = mat->GetName();
 		str.ToLower();
-		if (str.Icmpn("lights/", strlen("lights/")) == 0 || str.Icmpn("fogs/", strlen("fogs/")) == 0) {
+		if (str.Icmpn("lights/", (int)strlen("lights/")) == 0 || str.Icmpn("fogs/", (int)strlen("fogs/")) == 0) {
 			m_wndLights.AddString(mat->GetName());
 		}
 	}
